Fixed a memory leak that occurred when we malloc a new file structure.
This file struct was actually not needed (wasn't used either). It probably didn't work because of the race codition when we open the file. Wrap it around semas and it works fine.

Also, removed the exec_value int from the thread declaration. It was superceded by the exit status system.